implementation without tests

+++ src/main/java/it/systems/paymentreviser/dto/PaymentDTO.java
@@ -5,5 +5,5 @@ import it.systems.paymentreviser.enums.ValidCurrency;
 
 import java.math.BigDecimal;
 
-public record PaymentDTO(long id, BigDecimal amount, ValidCurrency currency, PaymentType paymentType) {
+public record PaymentDTO(Integer id, BigDecimal amount, ValidCurrency currency, PaymentType paymentType) {
 }
